Update DirectoryCleanup.java
Now has three mandatory methods, instead of 2 mandatory + 1 redundant. IdentifyFiles now loops through an array of File objects, checks to make sure they contain a specified Prefix in their filename, and then evaluates their age. If they are found to be older than the designated cutoff date, they're added into another array of files that are marked for deletion. They are then passed into a final method that iterates through that array and deletes each file, given that the array is not empty.

This commit also contains backup code that has been commented out and will not be present in dev/production.